requirements.txt
pyslack~=0.5.0
python-dotenv~=1.0.0